refactor(nodes-base): Move Postgres date type parsers to a shared util
