+++ packages/nodes-base/nodes/Postgres/transport/index.ts
@@ -1,6 +1,4 @@
 import { formatPemBlock } from '@n8n/utils/format-pem-block';
-import { ConnectionPoolManager } from '@utils/connection-pool-manager';
-import { LOCALHOST } from '@utils/constants';
 import type {
 	IExecuteFunctions,
 	ICredentialTestFunctions,
@@ -11,9 +9,12 @@ import type {
 import { createServer, type AddressInfo, type Server } from 'node:net';
 import pgPromise from 'pg-promise';
 
+import { ConnectionPoolManager } from '@utils/connection-pool-manager';
+import { LOCALHOST } from '@utils/constants';
+import { getDateAsStringTypeParsers, parseDateToISO } from '@utils/postgres';
+
 import type {
 	ConnectionsData,
-	PgpClient,
 	PgpConnectionParameters,
 	PostgresNodeCredentials,
 	PostgresNodeOptions,
@@ -22,11 +23,6 @@ import type {
 // dataTypeIDs for bigint (int8) and numeric types in PostgreSQL
 const BIGINT_TYPE_ID = 20;
 const NUMERIC_TYPE_ID = 1700;
-
-// Array type OIDs, not exposed via pg-types `builtins`
-const DATE_ARRAY_TYPE_ID = 1182; // _date
-const TIMESTAMP_ARRAY_TYPE_ID = 1115; // _timestamp
-const TIMESTAMPTZ_ARRAY_TYPE_ID = 1185; // _timestamptz
 
 export function applyLargeNumbersReceive(e: {
 	data: Array<Record<string, unknown>>;
@@ -54,62 +50,6 @@ export function createReceiveHandler(
 	return (e: unknown) => {
 		if (largeNumbersOutput !== 'numbers') return;
 		applyLargeNumbersReceive(e as Parameters<typeof applyLargeNumbersReceive>[0]);
-	};
-}
-
-export function parseDateToISO(value: string) {
-	const parsedDate = new Date(value);
-
-	if (isNaN(parsedDate.getTime())) {
-		return value;
-	}
-	return parsedDate.toISOString();
-}
-
-// pg-promise's bundled pg-types exposes `arrayParser.create(source, transform)`,
-// but the published typings type `arrayParser` as a plain function.
-type ArrayParser = {
-	create: (source: string, transform: (entry: string) => unknown) => { parse: () => unknown[] };
-};
-
-/**
- * Per-connection type parsers that return date/timestamp columns (and their array
- * variants) as JSON-safe strings, so node output never contains live `Date` objects.
- * DATE stays as its `YYYY-MM-DD` wire value; timestamps are normalized to ISO. Every
- * other OID falls back to the default parser. Applied via the pg `types` connection
- * option so we don't mutate global pg-types state shared across all pools.
- */
-export function getDateAsStringTypeParsers(
-	pgp: PgpClient,
-): NonNullable<PgpConnectionParameters['types']> {
-	const { builtins } = pgp.pg.types;
-	const arrayParser = pgp.pg.types.arrayParser as unknown as ArrayParser;
-	const parseElements = (transform: (entry: string) => unknown) => (value: string) =>
-		arrayParser.create(value, transform).parse();
-
-	const overrides = new Map<number, (value: string) => unknown>([
-		[builtins.DATE, (value) => value],
-		[builtins.TIMESTAMP, parseDateToISO],
-		[builtins.TIMESTAMPTZ, parseDateToISO],
-		[DATE_ARRAY_TYPE_ID, parseElements((value) => value)],
-		[TIMESTAMP_ARRAY_TYPE_ID, parseElements(parseDateToISO)],
-		[TIMESTAMPTZ_ARRAY_TYPE_ID, parseElements(parseDateToISO)],
-	]);
-
-	return {
-		setTypeParser(
-			id: number,
-			formatOrParseFn: string | ((value: string) => unknown),
-			parseFn?: string | ((value: string) => unknown),
-		) {
-			const fn = parseFn ?? formatOrParseFn;
-			if (typeof fn === 'function') overrides.set(id, fn);
-		},
-		getTypeParser(id: number, format?: 'text' | 'binary') {
-			const override = overrides.get(id);
-			if (override && (format === undefined || format === 'text')) return override;
-			return pgp.pg.types.getTypeParser(id, format);
-		},
 	};
 }
 
@@ -191,11 +131,10 @@ export async function configurePostgres(
 		const dbConfig = getPostgresConfig(credentials, options);
 
 		if (typeof options.nodeVersion === 'number' && options.nodeVersion >= 2.7) {
-			// Return DATE, TIME and array date/timestamp columns as strings too, via a
-			// per-connection type config that leaves global pg-types state untouched.
+			// Also return DATE and date/timestamp array columns as strings
 			dbConfig.types = getDateAsStringTypeParsers(pgp);
 		} else if (typeof options.nodeVersion === 'number' && options.nodeVersion >= 2.1) {
-			// Return TIMESTAMP columns as ISO strings (DATE columns still return Date objects)
+			// DATE columns still return Date objects on these versions
 			[pgp.pg.types.builtins.TIMESTAMP, pgp.pg.types.builtins.TIMESTAMPTZ].forEach((type) => {
 				pgp.pg.types.setTypeParser(type, parseDateToISO);
 			});
